Make changes for presentation
We rework the tests. Drop unnecessary tests for the presentation.

Part of this includes:

Set up wallaby
---------------

LiveView setup is more complicated. We follow instructions in:

(search for LiveView in the setup
https://hexdocs.pm/wallaby/readme.html#setup)

We also update the AccountManagementWallabyTest ->
TimelinePostsWallabyTest. We don't need to show wallaby testing live and
dead views. Dead views are assumed.